SAVE SYSTEMMMMMMMMM + enemy health

+++ Assets/Scripts/Player/PlayerController.cs
@@ -20,7 +20,7 @@ namespace DeadLink.Player
         {
             public List<string> PlayerPowerUps;
 
-            public int HealthPoints;
+            public float HealthPoints;
             public int HealthBarCount;
             
             public int LastCheckPoint;
